avanochi/azure/AZ_functions/requirements.txt
azure-functions
azure-cosmos
python-dotenv
pyjwt